Update 6-02.csv
Removed the entries relating to PM size cut-off from here. This is relevant to the sampling procedure. The separation of sampling procedure and sample treatment may seem artificial, but I think we can keep it. This code table here is more relevant for grab samples, the sampling procedure code table is more relevant to online (aerosol) sampling.