Issue112 (#324)
* Create 6-02.csv

* Update 6-02.csv

Removed the entries relating to PM size cut-off from here. This is relevant to the sampling procedure. The separation of sampling procedure and sample treatment may seem artificial, but I think we can keep it. This code table here is more relevant for grab samples, the sampling procedure code table is more relevant to online (aerosol) sampling.

* Update 6-02.csv

* Update 6-02.csv

* corrected typo on "particle removal"

* Update 6-02.csv

* Update 6-02.csv

Co-authored-by: LuisaIckes <46481310+LuisaIckes@users.noreply.github.com>
Co-authored-by: Jörg Klausen <34302947+joergklausen@users.noreply.github.com>
Co-authored-by: jbianchi81 <84208498+jbianchi81@users.noreply.github.com>
Co-authored-by: Franziska Stürzl <72029328+fstuerzl@users.noreply.github.com>